+++ m8rscript/Function.h
@@ -70,7 +70,7 @@ public:
     int32_t addLocal(const Atom& name);
     int32_t localIndex(const Atom& name) const;
     Atom localName(int32_t index) const { return (index < static_cast<int32_t>(_locals.size())) ? _locals[index] : Atom(); }
-    size_t localSize() const { return _locals.size() + _tempRegisterCount; }
+    uint32_t localSize() const { return static_cast<uint32_t>(_locals.size()) + _tempRegisterCount; }
     
     virtual CallReturnValue call(ExecutionUnit*, Value thisValue, uint32_t nparams) override;
